Do not use variable tag as function
